requirements.txt
streamlit
PyPDF2
langchain
langchain-openai
tavily-python
python-dotenv
google-genai

